// ht-web-interface/src/models/account/AccountEvent.ts
export enum AccountEvent {
    Name = "account",
    Show = "account/SHOW",
    ShowSuccess = "account/SHOW_SUCCESS",
    ShowFailure = "account/SHOW_FAILURE",
    Edit = "account/EDIT",
    EditSuccess = "account/EDIT_SUCCESS",
    EditFailure = "account/EDIT_FAILURE",
    EditAccountSuccessReset = "account/EDIT_ACCOUNT_SUCCESS_RESET",
    ResetAccountError = "account/RESET_ACCOUNT_ERROR",
}